feat: Payment Index UI
